Add the files for users
This branch contains the html files for user sign in page/sign up page and user Profile page, as well as their respective
css Files. It also contains a background picture for the sign in/sign up pages.

User Sign in page- Any user should be able to sign in to his or her profile.
User Sign up Page- A new user should be able to register though the web app.
User Profile page- This is the page that generates information to a signed in user,
whereby
User Red-flag- signifies list of red-flag records a user has created and,
User intervention- signifies list of intervention records a user has created.